M051: Update M051 Series BSP CMSIS to V3.01.004
used commands:
cd ~/cheali-charger/src/hardware/nuvoton-M051/cpu/CMSIS
rm -rf CMSIS Device StdDriver
cd /tmp
unzip M051_Series_BSP_CMSIS_V3.01.004.zip
cd Library
cp -R CMSIS/ Device/ StdDriver/ ~/cheali-charger/src/hardware/nuvoton-M051/cpu/CMSIS

+++ src/hardware/nuvoton-M051/cpu/CMSIS/StdDriver/src/sys.c
@@ -1,31 +1,41 @@
 /**************************************************************************//**
  * @file     sys.c
  * @version  V3.00
- * $Revision: 6 $
- * $Date: 14/01/28 6:58p $
+ * $Revision: 12 $
+ * $Date: 15/05/20 2:07p $
  * @brief    M051 series SYS driver source file
  *
  * @note
+ * SPDX-License-Identifier: Apache-2.0
+ *
  * Copyright (C) 2013 Nuvoton Technology Corp. All rights reserved.
 *****************************************************************************/
 #include "M051Series.h"
-/** @addtogroup M051_Device_Driver M051 Device Driver
+/** @addtogroup Standard_Driver Standard Driver
   @{
 */
 
-/** @addtogroup M051_SYS_Driver SYS Driver
+/** @addtogroup SYS_Driver SYS Driver
   @{
 */
 
 
-/** @addtogroup M051_SYS_EXPORTED_FUNCTIONS SYS Exported Functions
+/** @addtogroup SYS_EXPORTED_FUNCTIONS SYS Exported Functions
   @{
 */
 
 /**
-  * @brief  This function clear the selected system reset source
-  * @param  u32Src is system reset source
-  * @return None
+  * @brief      Clear reset source 
+  * @param[in]  u32Src is system reset source. Including :
+  *             - \ref SYS_RSTSRC_RSTS_CPU_Msk
+  *             - \ref SYS_RSTSRC_RSTS_MCU_Msk
+  *             - \ref SYS_RSTSRC_RSTS_BOD_Msk
+  *             - \ref SYS_RSTSRC_RSTS_LVR_Msk
+  *             - \ref SYS_RSTSRC_RSTS_WDT_Msk 
+  *             - \ref SYS_RSTSRC_RSTS_RESET_Msk 
+  *             - \ref SYS_RSTSRC_RSTS_POR_Msk  
+  * @return     None
+  * @details    This function clear the selected system reset source.    
   */
 void SYS_ClearResetSrc(uint32_t u32Src)
 {
@@ -33,10 +43,12 @@ void SYS_ClearResetSrc(uint32_t u32Src)
 }
 
 /**
-  * @brief  This function get Brown-out detector output status
-  * @return 0: System voltage is higher than BOD_VL setting or BOD_EN is 0.
-  *         1: System voltage is lower than BOD_VL setting.
-  *         Note : If the BOD_EN is 0, this function always return 0.
+  * @brief      Get Brown-out detector output status
+  * @param      None
+  * @retval     0 System voltage is higher than BOD_VL setting or BOD_EN is 0.
+  * @retval     1 System voltage is lower than BOD_VL setting.
+  * @details    This function get Brown-out detector output status.
+  *             If the BOD_EN is 0, this function always return 0.  
   */
 uint32_t SYS_GetBODStatus(void)
 {
@@ -44,8 +56,10 @@ uint32_t SYS_GetBODStatus(void)
 }
 
 /**
-  * @brief  This function get the system reset source register value
-  * @return Reset source
+  * @brief      Get reset source
+  * @param      None
+  * @return     Reset source
+  * @details    This function get the system reset source register value.  
   */
 uint32_t SYS_GetResetSrc(void)
 {
@@ -53,9 +67,11 @@ uint32_t SYS_GetResetSrc(void)
 }
 
 /**
-  * @brief  This function check register write-protection bit setting
-  * @return 0: Write-protection function is disabled.
-  *         1: Write-protection function is enabled.
+  * @brief      Check if register lock is set
+  * @param      None
+  * @retval     0 Write-protection function is disabled.
+  * @retval     1 Write-protection function is enabled.
+  * @details    This function check register write-protection bit setting.  
   */
 uint32_t SYS_IsRegLocked(void)
 {
@@ -63,8 +79,10 @@ uint32_t SYS_IsRegLocked(void)
 }
 
 /**
-  * @brief  This function get product ID.
-  * @return Product ID
+  * @brief      Get product ID
+  * @param      None
+  * @return     Product ID
+  * @details    This function get product ID.  
   */
 uint32_t  SYS_ReadPDID(void)
 {
@@ -72,8 +90,10 @@ uint32_t  SYS_ReadPDID(void)
 }
 
 /**
-  * @brief  This function reset chip.
-  * @return None
+  * @brief      Reset chip with chip reset
+  * @param      None
+  * @return     None
+  * @details    This function reset chip with chip reset.  
   */
 void SYS_ResetChip(void)
 {
@@ -81,8 +101,10 @@ void SYS_ResetChip(void)
 }
 
 /**
-  * @brief  This function reset CPU.
-  * @return None
+  * @brief      Reset chip with CPU reset
+  * @param      None
+  * @return     None
+  * @details    This function reset CPU with CPU reset.  
   */
 void SYS_ResetCPU(void)
 {
@@ -90,49 +112,52 @@ void SYS_ResetCPU(void)
 }
 
 /**
-  * @brief  This function reset selected modules.
-  * @param  u32ModuleIndex is module index. Including :
-  *           - \ref CHIP_RST
-  *           - \ref CPU_RST
-  *           - \ref EBI_RST
-  *           - \ref HDIV_RST
-  *           - \ref GPIO_RST
-  *           - \ref TMR0_RST
-  *           - \ref TMR1_RST
-  *           - \ref TMR2_RST
-  *           - \ref TMR3_RST
-  *           - \ref I2C0_RST
-  *           - \ref I2C1_RST
-  *           - \ref SPI0_RST
-  *           - \ref SPI1_RST
-  *           - \ref UART0_RST
-  *           - \ref UART1_RST
-  *           - \ref PWM03_RST
-  *           - \ref PWM47_RST
-  *           - \ref ACMP01_RST
-  *           - \ref ACMP23_RST
-  *           - \ref ADC_RST  
-  * @return None
+  * @brief      Reset Module
+  * @param[in]  u32ModuleIndex is module index. Including :
+  *             - \ref EBI_RST
+  *             - \ref HDIV_RST
+  *             - \ref GPIO_RST
+  *             - \ref TMR0_RST
+  *             - \ref TMR1_RST
+  *             - \ref TMR2_RST
+  *             - \ref TMR3_RST
+  *             - \ref I2C0_RST
+  *             - \ref I2C1_RST
+  *             - \ref SPI0_RST
+  *             - \ref SPI1_RST
+  *             - \ref UART0_RST
+  *             - \ref UART1_RST
+  *             - \ref PWM03_RST
+  *             - \ref PWM47_RST
+  *             - \ref ACMP01_RST
+  *             - \ref ACMP23_RST
+  *             - \ref ADC_RST
+  * @return     None
+  * @details    This function reset selected module.    
   */
 void SYS_ResetModule(uint32_t u32ModuleIndex)
 {
-    *(volatile uint32_t *)(&(SYS->IPRSTC1) + (u32ModuleIndex >> 24)) |= 1 << (u32ModuleIndex & 0x00ffffff);
-    *(volatile uint32_t *)(&(SYS->IPRSTC1) + (u32ModuleIndex >> 24)) &= ~(1 << (u32ModuleIndex & 0x00ffffff));
+    /* Generate reset signal to the corresponding module */
+    *(volatile uint32_t *)((uint32_t)&SYS->IPRSTC1 + (u32ModuleIndex >> 24))  |= 1 << (u32ModuleIndex & 0x00ffffff);
+
+    /* Release corresponding module from reset state */
+    *(volatile uint32_t *)((uint32_t)&SYS->IPRSTC1 + (u32ModuleIndex >> 24))  &= ~(1 << (u32ModuleIndex & 0x00ffffff));
 }
 
 /**
-  * @brief  This function configure BOD function.
-  *         Configure BOD reset or interrupt mode and set Brown-out voltage level.
-  *         Enable Brown-out function
-  * @param  i32Mode is reset or interrupt mode. Including :
-  *         - \ref SYS_BODCR_BOD_RST_EN
-  *         - \ref SYS_BODCR_BOD_INTERRUPT_EN
-  * @param  u32BODLevel is Brown-out voltage level. Including :
-  *         - \ref SYS_BODCR_BOD_VL_4_4V
-  *         - \ref SYS_BODCR_BOD_VL_3_7V
-  *         - \ref SYS_BODCR_BOD_VL_2_7V
-  *         - \ref SYS_BODCR_BOD_VL_2_2V
-  * @return None
+  * @brief      Enable and set Brown-out detector function
+  * @param[in]  i32Mode is reset or interrupt mode. Including :
+  *             - \ref SYS_BODCR_BOD_RST_EN
+  *             - \ref SYS_BODCR_BOD_INTERRUPT_EN
+  * @param[in]  u32BODLevel is Brown-out voltage level. Including :
+  *             - \ref SYS_BODCR_BOD_VL_4_4V
+  *             - \ref SYS_BODCR_BOD_VL_3_7V
+  *             - \ref SYS_BODCR_BOD_VL_2_7V
+  *             - \ref SYS_BODCR_BOD_VL_2_2V
+  * @return     None
+  * @details    This function configure Brown-out detector function.
+  *             It configure Brown-out detector reset or interrupt mode, enable Brown-out function and set Brown-out voltage level.
+  *             
   */
 void SYS_EnableBOD(int32_t i32Mode, uint32_t u32BODLevel)
 {
@@ -142,8 +167,10 @@ void SYS_EnableBOD(int32_t i32Mode, uint32_t u32BODLevel)
 }
 
 /**
-  * @brief  This function disable BOD function.
-  * @return None
+  * @brief      Disable Brown-out detector function
+  * @param      None
+  * @return     None
+  * @details    This function disable Brown-out detector function.   
   */
 void SYS_DisableBOD(void)
 {
@@ -152,10 +179,10 @@ void SYS_DisableBOD(void)
 
 
 
-/*@}*/ /* end of group M051_SYS_EXPORTED_FUNCTIONS */
+/*@}*/ /* end of group SYS_EXPORTED_FUNCTIONS */
 
-/*@}*/ /* end of group M051_SYS_Driver */
+/*@}*/ /* end of group SYS_Driver */
 
-/*@}*/ /* end of group M051_Device_Driver */
+/*@}*/ /* end of group Standard_Driver */
 
 /*** (C) COPYRIGHT 2013 Nuvoton Technology Corp. ***/
